requirements.txt
langchain>=0.0.10
langchain-openai>=0.0.1
langchain-community>=0.0.10
openai>=1.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
flask>=2.0.0
flask-sqlalchemy>=3.0.0
stripe>=8.0.0
pinecone-client>=2.2.4
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
textual>=0.1.18
sqlalchemy>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
httpx>=0.24.0
python-multipart>=0.0.5 